refactor: consolidate test utilities into test_helpers_test.go and replace inline implementations
